Improve error message rendering
